backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
scikit-learn
httpx
python-multipart
python-dotenv
firebase-admin
twilio
googlemaps
newsapi-python
